requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==20.8b1
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
Pygments==2.7.4
python-dotenv==0.15.0
regex==2020.11.13
SQLAlchemy==1.3.23
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
